Add ActionLoadDetail and ActionLoadDetailDetail
Features a complete emulation of the Action loading system.
Verified to be accurate enough to play Extensions -- if problems
come up with it, raise an issue.

These two tags are me basically just going absolutely nuts with
debugging the ConditionalAction parser due to an issue
A Casual Guy#0070 came to me with during the development of
Hacknet: Flagged.